LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-page/iframe-body-margin-attributes-expected.txt



FAIL body marginwidth vs body leftmargin assert_equals: marginTop expected "20px" but got "10px"
FAIL iframe marginwidth vs child body leftmargin assert_equals: marginTop expected "20px" but got "10px"
FAIL iframe marginwidth vs child body marginwidth assert_equals: marginTop expected "20px" but got "10px"
